apps/uscis-fill-local/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "uscis-fill"
version = "0.1.0"
description = "Local-first Tally intake import, review, and USCIS PDF draft fill"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=3.1",
    "sqlalchemy>=2.0",
    "pypdf>=4.0",
    "pyyaml>=6",
    "streamlit>=1.28",
]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.1"]

[project.scripts]
uscis-fill = "uscis_fill.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]
